organize utils into own folder
